requirements/base.txt
Flask==2.2.2 
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-User==1.0.2.2
Flask-Admin==1.6.0
Flask-API==3.0.post1
email_validator==1.2.1
python-dotenv==0.20.0
PyJWT==2.4.0
pytest==7.1.2